refactoring: ComponentsStyle.js 의 내부 컴포넌트의 파일 분리
EmptyContentBox 컴포넌트 파일 분리
titleBox 컴포넌트 파일분리
ColumnFooterStyledComponent 파일 분리 및 컴포넌트 압축
