+++ frontend/host-app/src/components/ColumnFooter.js
@@ -1,24 +1,32 @@
 import React from "react";
 import {Icon} from "@material-ui/core";
-import {FooterBox, FooterStyle} from "./ComponentsStyle";
 import useStyles from "./Questions/useStyles";
+import ColumnFooterStyledComponent from "./StyledComponent/ColumnFooterStyledComponent.js";
 
 function ColumnFooter({data, handler}) {
 	const classes = useStyles();
 	const increaseHeight = () => handler(data + 1);
-	const decreaseHeight = () => ((data > 0) && handler(data - 1));
+	const decreaseHeight = () => data > 0 && handler(data - 1);
 
 	return (
-		<FooterStyle>
-			<FooterBox>
-				<Icon className={classes.footerButton} onClick={() => { decreaseHeight(); }}>
-					remove
-				</Icon>
-				<Icon className={classes.footerButton} onClick={() => { increaseHeight(); }}>
-					add
-				</Icon>
-			</FooterBox>
-		</FooterStyle>
+		<ColumnFooterStyledComponent>
+			<Icon
+				className={classes.footerButton}
+				onClick={() => {
+					decreaseHeight();
+				}}
+			>
+				remove
+			</Icon>
+			<Icon
+				className={classes.footerButton}
+				onClick={() => {
+					increaseHeight();
+				}}
+			>
+				add
+			</Icon>
+		</ColumnFooterStyledComponent>
 	);
 }
