+++ frontend/host-app/src/components/StyledComponent/EmptyContentBox.js
@@ -1,0 +1,17 @@
+import styled from "styled-components";
+
+const EmptyContentBox = styled.div`
+	display: flex;
+	flex-direction: column;
+	// flex: 1;
+	align-items: center;
+	// overflow: auto;
+	// border-radius: 8px;
+	background-color: #f1f3f5;
+	border: 1px solid #e9ecef;
+	// height: 100%;
+	box-sizing: border-box;
+	margin-top: -1rem;
+`;
+
+export default EmptyContentBox;
