+++ frontend/host-app/src/components/StyledComponent/TitleBox.js
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+const TitleBox = styled.div`
+	display: flex;
+	align-items: center;
+	width: 100%;
+	justify-content: space-around;
+	min-height: 2.5rem;
+`;
+
+export default TitleBox;
