+++ frontend/host-app/src/components/SwitchTitle.js
@@ -4,7 +4,8 @@ import Badge from "@material-ui/core/Badge";
 import {makeStyles} from "@material-ui/core";
 import {socketClient, useSocket} from "../libs/socket.io-Client-wrapper";
 import {HostContext} from "../libs/hostContext";
-import {TitleStyle, TitleBox} from "./ComponentsStyle";
+import {TitleStyle} from "./StyledComponent/ComponentsStyle.js";
+import TitleBox from "./StyledComponent/TitleBox.js";
 
 const useStyles = makeStyles(theme => ({
 	margin: {
